+++ lib/mounter/http-face.d.ts
@@ -1,0 +1,9 @@
+/**
+ * @param {string} methods
+ * @param {import('../../index.js').FaceOption} face
+ * @param {import('../../index.js').Mare[]} maresBefore
+ * @param {import('../../index.js').Mare[]} maresAfter
+ * @param {string} facePrefix
+ * @param {import('../../index.js').DesireWithHarbour} desire
+ */
+export default function mountHTTPFace(methods: string, face: import('../../index.js').FaceOption, maresBefore: import('../../index.js').Mare[], maresAfter: import('../../index.js').Mare[], facePrefix: string, desire: import('../../index.js').DesireWithHarbour): Promise<void>;