+++ lib/mounter/http-fold.d.ts
@@ -1,0 +1,5 @@
+/**
+ * @param {import('../../index.js').FolderOption} fold
+ * @param {import('../../index.js').DesireWithHarbour} desire
+ */
+export default function mountHTTPFolder(fold: import('../../index.js').FolderOption, desire: import('../../index.js').DesireWithHarbour): Promise<void>;