+++ lib/mounter/wock-face.d.ts
@@ -1,0 +1,8 @@
+/**
+ * @param {import('../../index.js').FaceOption} face
+ * @param {import('../../index.js').Mare[]} maresBefore
+ * @param {import('../../index.js').Mare[]} maresAfter
+ * @param {import('../../index.js').DesireWithHarbour} desire
+ * @returns {Promise<void>}
+ */
+export default function mountWockFace(face: import('../../index.js').FaceOption, maresBefore: import('../../index.js').Mare[], maresAfter: import('../../index.js').Mare[], desire: import('../../index.js').DesireWithHarbour): Promise<void>;